created a PostForm component and moved related code out of dashboard route to it
